refactor: move forms and add typings
M2-546

refactor: re-add component names

refactor: remove unnecessary whitespace

refactor: remove unnecessary whitespace

chore: remove now empty profile update form file
