+++ packages/theme/modules/customer/index.ts
@@ -18,8 +18,7 @@ const moduleRoutes : NuxtRouteConfig = {
     {
       name: 'customer-my-profile',
       path: 'my-profile',
-      // component: path.resolve(moduleDir, 'pages/MyProfile.vue'),
-      component: stub,
+      component: path.resolve(moduleDir, 'pages/MyProfile/MyProfile.vue'),
       meta: {
         titleLabel: 'My profile',
       },
