+++ packages/theme/modules/customer/pages/MyAccount/MyAccount.vue
@@ -62,6 +62,7 @@ import {
 import { useSidebarLinkGroups } from './useSidebarLinkGroups';
 
 export default defineComponent({
+  name: 'MyAccount',
   components: {
     SfBar,
     SfList,
